[#25274] MPI pairing dispatch correction

+++ bibfor/cont_pair/aprema.F90
@@ -132,10 +132,10 @@ implicit none
         if(one_proc)then
             nb_proc = 1
         endif
-        nb_poin_mpi  = nb_poin/nb_proc
+        nb_poin_mpi  = int(nb_poin/nb_proc)
         nbr_poin_mpi = nb_poin-nb_poin_mpi*nb_proc
         idx_start   = 1+(i_proc)*nb_poin_mpi
-        idx_end     = idx_start+nb_poin_mpi-1+(nbr_poin_mpi*(i_proc+1)/nb_proc)
+        idx_end     = idx_start+nb_poin_mpi-1+(nbr_poin_mpi*int((i_proc+1)/nb_proc))
 
 !
 ! ----- Loop on points
